ofz#820 oom terminate
comments claim return is len of data, but negative numbers for compressed data errors
are returned. Make things return the amount of successful data read and turn into
unsigned size_t to match ofz#727

+++ hwpfilter/source/hgzip.h
@@ -80,7 +80,7 @@ int     gz_close    ( gz_stream *file );
  * @param len Length of data to be read
  * @returns The number of bytes actually read
  */
-int     gz_read     ( gz_stream *file, voidp  buf, unsigned len );
+size_t     gz_read     ( gz_stream *file, voidp  buf, unsigned len );
 /**
  * Flushes all pending output into the compressed file
  * gz_flush should be called only when strictly necessary because it can
